refector: ViewService의 기능을 InputView와 OutputView로 각각 분리
