+++ docs/README.md
@@ -54,4 +54,4 @@
 
 ### 기타
 
-- [ ] ViewService의 기능을 InputView와 OutputView로 각각 분리한다.
+- [x] ViewService의 기능을 InputView와 OutputView로 각각 분리한다.
