+++ src/main/java/baseball/controller/GameController.java
@@ -4,17 +4,20 @@ import baseball.application.NumberService;
 import baseball.domain.Hint;
 import baseball.domain.Player;
 import baseball.util.NumberUtil;
-import baseball.view.ViewService;
+import baseball.view.InputView;
+import baseball.view.OutputView;
 import java.util.Objects;
 
 public class GameController {
 
     private final NumberService numberService;
-    private final ViewService viewService;
+    private final InputView inputView;
+    private final OutputView outputView;
 
-    public GameController(NumberService numberService, ViewService viewService) {
+    public GameController(NumberService numberService, InputView inputView, OutputView outputView) {
         this.numberService = numberService;
-        this.viewService = viewService;
+        this.inputView = inputView;
+        this.outputView = outputView;
     }
 
     public void startGame() {
@@ -23,7 +26,7 @@ public class GameController {
 
             play(computer);
 
-        } while (!Objects.equals(viewService.readNumber(), "2")); // 종료(2)를 누르지 않으면 게임을 다시 시작한다.
+        } while (!Objects.equals(inputView.readNumber(), "2")); // 종료(2)를 누르지 않으면 게임을 다시 시작한다.
     }
 
     private void play(Player computer) {
@@ -31,7 +34,7 @@ public class GameController {
             Player player = preparePlayer();
 
             Hint hint = numberService.getHint(computer, player); // 컴퓨터와 플레이어의 숫자를 비교하여 힌트를 얻는다.
-            viewService.printHintMessage(hint); // 힌트를 출력, 만약 숫자가 맞는다면 게임 종료 메시지도 출력한다.
+            outputView.printHintMessage(hint); // 힌트를 출력, 만약 숫자가 맞는다면 게임 종료 메시지도 출력한다.
             if (hint.isThreeStrike()) { // 숫자를 맞췄다면 끝낸다.
                 break;
             }
@@ -47,8 +50,8 @@ public class GameController {
     private Player preparePlayer() {
         Player player = new Player();
 
-        viewService.printReadMessage();
-        String number = viewService.readNumber();
+        outputView.printReadMessage();
+        String number = inputView.readNumber();
         NumberUtil.validate(number);
         player.setPlayer(NumberUtil.converStringToList(number));
