+++ src/main/java/baseball/view/InputView.java
@@ -1,0 +1,11 @@
+package baseball.view;
+
+import static camp.nextstep.edu.missionutils.Console.readLine;
+
+public class InputView {
+
+    public String readNumber() {
+        return readLine();
+    }
+
+}
